address Mayukha's comments: rename files and variables to match defs.json - oracle_set and oracle_delete
