+++ .ci-config/rippled.cfg
@@ -170,5 +170,11 @@ fixNFTokenRemint
 # 2.0.0-b4 Amendments
 XChainBridge
 DID
-# 2.2.0-b1 Amendments
+# 2.2.0-b3 Amendments
+fixNFTokenReserve
+fixInnerObjTemplate
+fixAMMOverflowOffer
 PriceOracle
+fixEmptyDID
+fixXChainRewardRounding
+fixPreviousTxnID
