+++ xrpl/models/transactions/types/transaction_type.py
@@ -18,7 +18,6 @@ class TransactionType(str, Enum):
     CHECK_CASH = "CheckCash"
     CHECK_CREATE = "CheckCreate"
     CLAWBACK = "Clawback"
-    DELETE_ORACLE = "OracleDelete"
     DEPOSIT_PREAUTH = "DepositPreauth"
     DID_DELETE = "DIDDelete"
     DID_SET = "DIDSet"
@@ -32,11 +31,12 @@ class TransactionType(str, Enum):
     NFTOKEN_MINT = "NFTokenMint"
     OFFER_CANCEL = "OfferCancel"
     OFFER_CREATE = "OfferCreate"
+    ORACLE_SET = "OracleSet"
+    ORACLE_DELETE = "OracleDelete"
     PAYMENT = "Payment"
     PAYMENT_CHANNEL_CLAIM = "PaymentChannelClaim"
     PAYMENT_CHANNEL_CREATE = "PaymentChannelCreate"
     PAYMENT_CHANNEL_FUND = "PaymentChannelFund"
-    SET_ORACLE = "OracleSet"
     SET_REGULAR_KEY = "SetRegularKey"
     SIGNER_LIST_SET = "SignerListSet"
     TICKET_CREATE = "TicketCreate"
